+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 1.0.0
+
+- [feature] Trap events in modals by default
+- [fix] Fix coloring values for popover
+- [fix] rename unsafe lifecycle methods to eliminate warnings under React 16
 
 ## 0.12.0